Allow ImageJ to run in headless mode
The IJ.run command creates requests to the interface, even if it is
not displayed. For such requests to work, a X server must be running on
Linux. This makes running the code remotely cumbersome and possibly slower.

Here we run the code winthin a fake X11 server (Xfvb).
TODO: need to check that it works in OS X.